added swagger to sql actix server. seperated swagger to shared generic module
