+++ apps/rust/actix_app/src/swagger.rs
@@ -1,23 +1,11 @@
 use crate::mongo::ErrorResponse;
-use actix_web::{
-    dev::{Service, ServiceRequest, ServiceResponse, Transform},
-    HttpResponse,
-};
-use futures::future::LocalBoxFuture;
-use std::future::{self, Ready};
-use utoipa::{
-    openapi::security::{ApiKey, ApiKeyValue, SecurityScheme},
-    Modify, OpenApi,
-};
-
+use utoipa::OpenApi;
+use rust_servers_shared::{SecurityAddon};
 use crate::product;
 use crate::todo;
 use crate::user;
 use rust_author_api::author;
 use rust_books_api::book;
-
-const API_KEY_NAME: &str = "todo_apikey";
-const API_KEY: &str = "utoipa-rocks";
 
 #[derive(OpenApi)]
 #[openapi(
@@ -74,138 +62,3 @@ tags(
 modifiers(&SecurityAddon)
 )]
 pub struct ApiDoc1;
-
-pub struct SecurityAddon;
-
-impl Modify for SecurityAddon {
-    fn modify(&self, openapi: &mut utoipa::openapi::OpenApi) {
-        let components = openapi.components.as_mut().unwrap(); // we can unwrap safely since there already is components registered.
-        components.add_security_scheme(
-            "api_key",
-            SecurityScheme::ApiKey(ApiKey::Header(ApiKeyValue::new("todo_apikey"))),
-        )
-    }
-}
-
-/// Require api key middlware will actually require valid api key
-pub struct RequireApiKey;
-
-impl<S> Transform<S, ServiceRequest> for RequireApiKey
-where
-    S: Service<
-        ServiceRequest,
-        Response = ServiceResponse<actix_web::body::BoxBody>,
-        Error = actix_web::Error,
-    >,
-    S::Future: 'static,
-{
-    type Response = ServiceResponse<actix_web::body::BoxBody>;
-    type Error = actix_web::Error;
-    type Transform = ApiKeyMiddleware<S>;
-    type InitError = ();
-    type Future = Ready<Result<Self::Transform, Self::InitError>>;
-
-    fn new_transform(&self, service: S) -> Self::Future {
-        future::ready(Ok(ApiKeyMiddleware {
-            service,
-            log_only: false,
-        }))
-    }
-}
-
-/// Log api key midleware only logs about missing or invalid api keys
-pub struct LogApiKey;
-
-impl<S> Transform<S, ServiceRequest> for LogApiKey
-where
-    S: Service<
-        ServiceRequest,
-        Response = ServiceResponse<actix_web::body::BoxBody>,
-        Error = actix_web::Error,
-    >,
-    S::Future: 'static,
-{
-    type Response = ServiceResponse<actix_web::body::BoxBody>;
-    type Error = actix_web::Error;
-    type Transform = ApiKeyMiddleware<S>;
-    type InitError = ();
-    type Future = Ready<Result<Self::Transform, Self::InitError>>;
-
-    fn new_transform(&self, service: S) -> Self::Future {
-        future::ready(Ok(ApiKeyMiddleware {
-            service,
-            log_only: true,
-        }))
-    }
-}
-
-pub struct ApiKeyMiddleware<S> {
-    service: S,
-    log_only: bool,
-}
-
-impl<S> Service<ServiceRequest> for ApiKeyMiddleware<S>
-where
-    S: Service<
-        ServiceRequest,
-        Response = ServiceResponse<actix_web::body::BoxBody>,
-        Error = actix_web::Error,
-    >,
-    S::Future: 'static,
-{
-    type Response = ServiceResponse<actix_web::body::BoxBody>;
-    type Error = actix_web::Error;
-    type Future = LocalBoxFuture<'static, Result<Self::Response, actix_web::Error>>;
-
-    fn poll_ready(
-        &self,
-        ctx: &mut core::task::Context<'_>,
-    ) -> std::task::Poll<Result<(), Self::Error>> {
-        self.service.poll_ready(ctx)
-    }
-
-    fn call(&self, req: ServiceRequest) -> Self::Future {
-        let response = |req: ServiceRequest, response: HttpResponse| -> Self::Future {
-            Box::pin(async { Ok(req.into_response(response)) })
-        };
-
-        match req.headers().get(API_KEY_NAME) {
-            Some(key) if key != API_KEY => {
-                if self.log_only {
-                    log::debug!("Incorrect api api provided!!!")
-                } else {
-                    return response(
-                        req,
-                        HttpResponse::Unauthorized().json(ErrorResponse::Unauthorized(
-                            String::from("incorrect api key"),
-                        )),
-                    );
-                }
-            }
-            None => {
-                if self.log_only {
-                    log::debug!("Missing api key!!!")
-                } else {
-                    return response(
-                        req,
-                        HttpResponse::Unauthorized()
-                            .json(ErrorResponse::Unauthorized(String::from("missing api key"))),
-                    );
-                }
-            }
-            _ => (), // just passthrough
-        }
-
-        if self.log_only {
-            log::debug!("Performing operation")
-        }
-
-        let future = self.service.call(req);
-
-        Box::pin(async move {
-            let response = future.await?;
-
-            Ok(response)
-        })
-    }
-}
