+++ apps/rust/actix_sql/Cargo.toml
@@ -5,13 +5,16 @@ edition = "2021"
 
 
 [dependencies]
+mongo = { path = "../../../libs/rust/mongo" }
 actix-web = { workspace = true }
 serde = { workspace = true }
+serde_json = { workspace = true }
 sqlx = { workspace = true }
 rust_servers_shared = { workspace = true }
 rust_proc_macros = { workspace = true }
 log = { workspace = true }
 utoipa = { workspace = true }
+futures = { workspace = true }
 utoipa-swagger-ui = { workspace = true }
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
