+++ apps/rust/actix_sql/src/books/model.rs
@@ -1,7 +1,8 @@
 use serde::{Deserialize, Serialize};
 use sqlx::FromRow;
+use utoipa::ToSchema;
 
-#[derive(FromRow, Deserialize, Serialize, Debug)]
+#[derive(ToSchema, FromRow, Deserialize, Serialize, Debug)]
 pub struct Book {
     pub id: i32,
     pub title: String,
@@ -9,14 +10,14 @@ pub struct Book {
     pub description: String,
 }
 
-#[derive(Deserialize, Serialize, Debug, Clone)]
+#[derive(ToSchema, Deserialize, Serialize, Debug, Clone)]
 pub struct CreateBook {
     pub title: String,
     pub author: String,
     pub description: String,
 }
 
-#[derive(Deserialize, Serialize, Debug, Clone)]
+#[derive(ToSchema, Deserialize, Serialize, Debug, Clone)]
 pub struct UpdateBook {
     pub title: String,
     pub author: String,
