+++ apps/rust/actix_sql/src/todos/mod.rs
@@ -3,3 +3,5 @@ mod handlers;
 mod model;
 
 pub use configure::*;
+pub use model::*;
+pub use handlers::*;
