+++ libs/rust/servers_shared/Cargo.toml
@@ -8,4 +8,8 @@ actix-web = { workspace = true }
 actix-cors = { workspace = true }
 env_logger = { workspace = true }
 sqlx = { workspace = true }
+utoipa = { workspace = true }
+futures = { workspace = true }
+log = { workspace = true }
+mongo = { path = "../mongo" }
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
